+++ CHANGES.md
@@ -14,6 +14,8 @@
 
 * base: add lakeshore module. by @guirodrigueslima in
   https://github.com/cnpem/epics-in-docker/pull/117
+* base: add motorParker. by @guirodrigueslima in
+  https://github.com/cnpem/epics-in-docker/pull/116
 * base: update areaDetector to R3-14. by @henriquesimoes in
   https://github.com/cnpem/epics-in-docker/pull/53